consolidate user/org/namespace navitems and routes
