+++ client/web/BUILD.bazel
@@ -988,10 +988,6 @@ ts_project(
         "src/enterprise/insights/stores/code-insights-license.ts",
         "src/enterprise/insights/stores/index.ts",
         "src/enterprise/main.tsx",
-        "src/enterprise/namespaces/navitems.ts",
-        "src/enterprise/namespaces/routes.ts",
-        "src/enterprise/organizations/navitems.ts",
-        "src/enterprise/organizations/routes.tsx",
         "src/enterprise/organizations/settings/routes.tsx",
         "src/enterprise/organizations/settings/sidebaritems.ts",
         "src/enterprise/own/DeleteFileButton.tsx",
@@ -1077,12 +1073,10 @@ ts_project(
         "src/enterprise/site-admin/productSubscription/ProductSubscriptionStatus.tsx",
         "src/enterprise/site-admin/productSubscription/SiteAdminProductSubscriptionPage.tsx",
         "src/enterprise/site-admin/routes.tsx",
-        "src/enterprise/user/navitems.ts",
         "src/enterprise/user/productSubscriptions/UserProductSubscriptionStatus.tsx",
         "src/enterprise/user/productSubscriptions/UserSubscriptionsProductSubscriptionPage.tsx",
         "src/enterprise/user/productSubscriptions/UserSubscriptionsProductSubscriptionsPage.tsx",
         "src/enterprise/user/productSubscriptions/backend.ts",
-        "src/enterprise/user/routes.tsx",
         "src/enterprise/user/settings/ExternalAccountNode.tsx",
         "src/enterprise/user/settings/UserEventLogsPage.tsx",
         "src/enterprise/user/settings/auth/UserSettingsPermissionsPage.tsx",
