+++ client/web/src/enterprise/EnterpriseWebApp.tsx
@@ -5,6 +5,8 @@ import '../SourcegraphWebApp.scss'
 import { logger } from '@sourcegraph/common'
 
 import { LegacySourcegraphWebApp } from '../LegacySourcegraphWebApp'
+import { orgAreaHeaderNavItems } from '../org/area/navitems'
+import { orgAreaRoutes } from '../org/area/routes'
 import { siteAdminSidebarGroups } from '../site-admin/sidebaritems'
 import { SourcegraphWebApp } from '../SourcegraphWebApp'
 import {
@@ -15,11 +17,11 @@ import {
 } from '../staticAppConfig'
 import type { AppShellInit } from '../storm/app-shell-init'
 import { routes } from '../storm/routes'
+import { userAreaHeaderNavItems } from '../user/area/navitems'
+import { userAreaRoutes } from '../user/area/routes'
 
 import { APP_ROUTES } from './app/routes'
 import { BrainDot } from './codeintel/dashboard/components/BrainDot'
-import { enterpriseOrgAreaHeaderNavItems } from './organizations/navitems'
-import { enterpriseOrganizationAreaRoutes } from './organizations/routes'
 import { enterpriseOrgSettingsAreaRoutes } from './organizations/settings/routes'
 import { enterpriseOrgSettingsSideBarItems } from './organizations/settings/sidebaritems'
 import { enterpriseRepoContainerRoutes } from './repo/enterpriseRepoContainerRoutes'
@@ -30,8 +32,6 @@ import { enterpriseRepoSettingsSidebarGroups } from './repo/settings/sidebaritem
 import { enterpriseRoutes } from './routes'
 import { enterpriseSiteAdminOverviewComponents } from './site-admin/overview/overviewComponents'
 import { enterpriseSiteAdminAreaRoutes } from './site-admin/routes'
-import { enterpriseUserAreaHeaderNavItems } from './user/navitems'
-import { enterpriseUserAreaRoutes } from './user/routes'
 import { enterpriseUserSettingsAreaRoutes } from './user/settings/routes'
 import { enterpriseUserSettingsSideBarItems } from './user/settings/sidebaritems'
 
@@ -42,14 +42,14 @@ const injectedValuesConfig = {
     siteAdminAreaRoutes: enterpriseSiteAdminAreaRoutes,
     siteAdminSideBarGroups: siteAdminSidebarGroups,
     siteAdminOverviewComponents: enterpriseSiteAdminOverviewComponents,
-    userAreaHeaderNavItems: enterpriseUserAreaHeaderNavItems,
-    userAreaRoutes: enterpriseUserAreaRoutes,
+    userAreaHeaderNavItems,
+    userAreaRoutes,
     userSettingsSideBarItems: enterpriseUserSettingsSideBarItems,
     userSettingsAreaRoutes: enterpriseUserSettingsAreaRoutes,
     orgSettingsSideBarItems: enterpriseOrgSettingsSideBarItems,
     orgSettingsAreaRoutes: enterpriseOrgSettingsAreaRoutes,
-    orgAreaRoutes: enterpriseOrganizationAreaRoutes,
-    orgAreaHeaderNavItems: enterpriseOrgAreaHeaderNavItems,
+    orgAreaRoutes,
+    orgAreaHeaderNavItems,
     repoContainerRoutes: enterpriseRepoContainerRoutes,
     repoRevisionContainerRoutes: enterpriseRepoRevisionContainerRoutes,
     repoHeaderActionButtons: enterpriseRepoHeaderActionButtons,
