--- client/web/src/enterprise/namespaces/navitems.ts
@@ -1,11 +1,0 @@
-import { BatchChangesIconNamespaceNav } from '../../batches/icons'
-import type { NamespaceAreaNavItem } from '../../namespaces/NamespaceArea'
-
-export const enterpriseNamespaceAreaHeaderNavItems: readonly NamespaceAreaNavItem[] = [
-    {
-        to: '/batch-changes',
-        label: 'Batch Changes',
-        icon: BatchChangesIconNamespaceNav,
-        condition: ({ batchChangesEnabled, isCodyApp }) => batchChangesEnabled && !isCodyApp,
-    },
-]
