--- client/web/src/enterprise/namespaces/routes.ts
@@ -1,4 +1,0 @@
-import type { NamespaceAreaRoute } from '../../namespaces/NamespaceArea'
-import { namespaceAreaRoutes } from '../../namespaces/routes'
-
-export const enterpriseNamespaceAreaRoutes: readonly NamespaceAreaRoute[] = [...namespaceAreaRoutes]
