--- client/web/src/enterprise/organizations/navitems.ts
@@ -1,8 +1,0 @@
-import { orgAreaHeaderNavItems } from '../../org/area/navitems'
-import type { OrgAreaHeaderNavItem } from '../../org/area/OrgHeader'
-import { enterpriseNamespaceAreaHeaderNavItems } from '../namespaces/navitems'
-
-export const enterpriseOrgAreaHeaderNavItems: readonly OrgAreaHeaderNavItem[] = [
-    ...orgAreaHeaderNavItems,
-    ...enterpriseNamespaceAreaHeaderNavItems,
-]
