--- client/web/src/enterprise/organizations/routes.tsx
@@ -1,58 +1,0 @@
-import { lazyComponent } from '@sourcegraph/shared/src/util/lazyComponent'
-
-import type { OrgAreaRoute } from '../../org/area/OrgArea'
-import { orgAreaRoutes } from '../../org/area/routes'
-import type { EditBatchSpecPageProps } from '../batches/batch-spec/edit/EditBatchSpecPage'
-import type { CreateBatchChangePageProps } from '../batches/create/CreateBatchChangePage'
-import type { NamespaceBatchChangesAreaProps } from '../batches/global/GlobalBatchChangesArea'
-import { enterpriseNamespaceAreaRoutes } from '../namespaces/routes'
-
-const NamespaceBatchChangesArea = lazyComponent<NamespaceBatchChangesAreaProps, 'NamespaceBatchChangesArea'>(
-    () => import('../batches/global/GlobalBatchChangesArea'),
-    'NamespaceBatchChangesArea'
-)
-
-const ExecuteBatchSpecPage = lazyComponent(
-    () => import('../batches/batch-spec/execute/ExecuteBatchSpecPage'),
-    'ExecuteBatchSpecPage'
-)
-
-const CreateBatchChangePage = lazyComponent<CreateBatchChangePageProps, 'CreateBatchChangePage'>(
-    () => import('../batches/create/CreateBatchChangePage'),
-    'CreateBatchChangePage'
-)
-
-const EditBatchSpecPage = lazyComponent<EditBatchSpecPageProps, 'EditBatchSpecPage'>(
-    () => import('../batches/batch-spec/edit/EditBatchSpecPage'),
-    'EditBatchSpecPage'
-)
-
-export const enterpriseOrganizationAreaRoutes: readonly OrgAreaRoute[] = [
-    ...orgAreaRoutes,
-    ...enterpriseNamespaceAreaRoutes,
-    {
-        path: 'batch-changes/create',
-        render: props => <CreateBatchChangePage headingElement="h1" {...props} initialNamespaceID={props.org.id} />,
-        condition: ({ batchChangesEnabled }) => batchChangesEnabled,
-        fullPage: true,
-    },
-    {
-        path: 'batch-changes/:batchChangeName/edit',
-        render: props => <EditBatchSpecPage {...props} />,
-        condition: ({ batchChangesEnabled, batchChangesExecutionEnabled }) =>
-            batchChangesEnabled && batchChangesExecutionEnabled,
-        fullPage: true,
-    },
-    {
-        path: 'batch-changes/:batchChangeName/executions/:batchSpecID/*',
-        render: props => <ExecuteBatchSpecPage {...props} />,
-        condition: ({ batchChangesEnabled, batchChangesExecutionEnabled }) =>
-            batchChangesEnabled && batchChangesExecutionEnabled,
-        fullPage: true,
-    },
-    {
-        path: 'batch-changes/*',
-        render: props => <NamespaceBatchChangesArea {...props} namespaceID={props.org.id} />,
-        condition: ({ batchChangesEnabled }) => batchChangesEnabled,
-    },
-]
