--- client/web/src/enterprise/user/navitems.ts
@@ -1,8 +1,0 @@
-import { userAreaHeaderNavItems } from '../../user/area/navitems'
-import type { UserAreaHeaderNavItem } from '../../user/area/UserAreaHeader'
-import { enterpriseNamespaceAreaHeaderNavItems } from '../namespaces/navitems'
-
-export const enterpriseUserAreaHeaderNavItems: readonly UserAreaHeaderNavItem[] = [
-    ...userAreaHeaderNavItems,
-    ...enterpriseNamespaceAreaHeaderNavItems,
-]
