+++ client/web/src/org/area/routes.tsx
@@ -2,9 +2,32 @@ import { Navigate } from 'react-router-dom'
 
 import { lazyComponent } from '@sourcegraph/shared/src/util/lazyComponent'
 
+import type { EditBatchSpecPageProps } from '../../enterprise/batches/batch-spec/edit/EditBatchSpecPage'
+import type { CreateBatchChangePageProps } from '../../enterprise/batches/create/CreateBatchChangePage'
+import type { NamespaceBatchChangesAreaProps } from '../../enterprise/batches/global/GlobalBatchChangesArea'
 import { namespaceAreaRoutes } from '../../namespaces/routes'
 
 import type { OrgAreaRoute } from './OrgArea'
+
+const NamespaceBatchChangesArea = lazyComponent<NamespaceBatchChangesAreaProps, 'NamespaceBatchChangesArea'>(
+    () => import('../../enterprise/batches/global/GlobalBatchChangesArea'),
+    'NamespaceBatchChangesArea'
+)
+
+const ExecuteBatchSpecPage = lazyComponent(
+    () => import('../../enterprise/batches/batch-spec/execute/ExecuteBatchSpecPage'),
+    'ExecuteBatchSpecPage'
+)
+
+const CreateBatchChangePage = lazyComponent<CreateBatchChangePageProps, 'CreateBatchChangePage'>(
+    () => import('../../enterprise/batches/create/CreateBatchChangePage'),
+    'CreateBatchChangePage'
+)
+
+const EditBatchSpecPage = lazyComponent<EditBatchSpecPageProps, 'EditBatchSpecPage'>(
+    () => import('../../enterprise/batches/batch-spec/edit/EditBatchSpecPage'),
+    'EditBatchSpecPage'
+)
 
 const OrgSettingsArea = lazyComponent(() => import('../settings/OrgSettingsArea'), 'OrgSettingsArea')
 
@@ -34,5 +57,31 @@ export const orgAreaRoutes: readonly OrgAreaRoute[] = [
     {
         path: 'account',
         render: () => <Navigate to="../settings/profile" />,
+    },
+
+    {
+        path: 'batch-changes/create',
+        render: props => <CreateBatchChangePage headingElement="h1" {...props} initialNamespaceID={props.org.id} />,
+        condition: ({ batchChangesEnabled }) => batchChangesEnabled,
+        fullPage: true,
+    },
+    {
+        path: 'batch-changes/:batchChangeName/edit',
+        render: props => <EditBatchSpecPage {...props} />,
+        condition: ({ batchChangesEnabled, batchChangesExecutionEnabled }) =>
+            batchChangesEnabled && batchChangesExecutionEnabled,
+        fullPage: true,
+    },
+    {
+        path: 'batch-changes/:batchChangeName/executions/:batchSpecID/*',
+        render: props => <ExecuteBatchSpecPage {...props} />,
+        condition: ({ batchChangesEnabled, batchChangesExecutionEnabled }) =>
+            batchChangesEnabled && batchChangesExecutionEnabled,
+        fullPage: true,
+    },
+    {
+        path: 'batch-changes/*',
+        render: props => <NamespaceBatchChangesArea {...props} namespaceID={props.org.id} />,
+        condition: ({ batchChangesEnabled }) => batchChangesEnabled,
     },
 ]
