+++ client/web/src/user/area/routes.tsx
@@ -2,9 +2,36 @@ import { Navigate } from 'react-router-dom'
 
 import { lazyComponent } from '@sourcegraph/shared/src/util/lazyComponent'
 
+import { RedirectRoute } from '../../components/RedirectRoute'
+import type { EditBatchSpecPageProps } from '../../enterprise/batches/batch-spec/edit/EditBatchSpecPage'
+import type { CreateBatchChangePageProps } from '../../enterprise/batches/create/CreateBatchChangePage'
+import type { NamespaceBatchChangesAreaProps } from '../../enterprise/batches/global/GlobalBatchChangesArea'
+import { SHOW_BUSINESS_FEATURES } from '../../enterprise/dotcom/productSubscriptions/features'
 import { namespaceAreaRoutes } from '../../namespaces/routes'
 
 import type { UserAreaRoute } from './UserArea'
+
+const AppSettingsArea = lazyComponent(() => import('../../enterprise/app/settings/AppSettingsArea'), 'AppSettingsArea')
+
+const NamespaceBatchChangesArea = lazyComponent<NamespaceBatchChangesAreaProps, 'NamespaceBatchChangesArea'>(
+    () => import('../../enterprise/batches/global/GlobalBatchChangesArea'),
+    'NamespaceBatchChangesArea'
+)
+
+const ExecuteBatchSpecPage = lazyComponent(
+    () => import('../../enterprise/batches/batch-spec/execute/ExecuteBatchSpecPage'),
+    'ExecuteBatchSpecPage'
+)
+
+const CreateBatchChangePage = lazyComponent<CreateBatchChangePageProps, 'CreateBatchChangePage'>(
+    () => import('../../enterprise/batches/create/CreateBatchChangePage'),
+    'CreateBatchChangePage'
+)
+
+const EditBatchSpecPage = lazyComponent<EditBatchSpecPageProps, 'EditBatchSpecPage'>(
+    () => import('../../enterprise/batches/batch-spec/edit/EditBatchSpecPage'),
+    'EditBatchSpecPage'
+)
 
 const UserSettingsArea = lazyComponent(() => import('../settings/UserSettingsArea'), 'UserSettingsArea')
 const UserProfile = lazyComponent(() => import('../profile/UserProfile'), 'UserProfile')
@@ -36,5 +63,51 @@ export const userAreaRoutes: readonly UserAreaRoute[] = [
     {
         path: 'account',
         render: () => <Navigate to="../profile" replace={true} />,
+    },
+
+    // Cody app specific route (cody/app settings page)
+    // This route won't be available for any non-app deploy types.
+    // See userAreaHeaderNavItems in client/web/src/enterprise/user/navitems.ts
+    // for more context on user settings page.
+    {
+        path: 'app-settings/*',
+        render: props => <AppSettingsArea telemetryService={props.telemetryService} />,
+        condition: context => context.isCodyApp,
+    },
+
+    // Redirect from previous /users/:username/subscriptions -> /users/:username/settings/subscriptions.
+    {
+        path: 'subscriptions/:page?',
+        render: () => (
+            <RedirectRoute
+                getRedirectURL={({ params }) => `../settings/subscriptions${params.page ? `/${params.page}` : ''}`}
+            />
+        ),
+        condition: () => SHOW_BUSINESS_FEATURES,
+    },
+    {
+        path: 'batch-changes/create',
+        render: props => <CreateBatchChangePage headingElement="h1" {...props} initialNamespaceID={props.user.id} />,
+        condition: ({ batchChangesEnabled }) => batchChangesEnabled,
+        fullPage: true,
+    },
+    {
+        path: 'batch-changes/:batchChangeName/edit',
+        render: props => <EditBatchSpecPage {...props} />,
+        condition: ({ batchChangesEnabled, batchChangesExecutionEnabled }) =>
+            batchChangesEnabled && batchChangesExecutionEnabled,
+        fullPage: true,
+    },
+    {
+        path: 'batch-changes/:batchChangeName/executions/:batchSpecID/*',
+        render: props => <ExecuteBatchSpecPage {...props} />,
+        condition: ({ batchChangesEnabled, batchChangesExecutionEnabled }) =>
+            batchChangesEnabled && batchChangesExecutionEnabled,
+        fullPage: true,
+    },
+    {
+        path: 'batch-changes/*',
+        render: props => <NamespaceBatchChangesArea {...props} namespaceID={props.user.id} />,
+        condition: ({ batchChangesEnabled }) => batchChangesEnabled,
     },
 ]
